requirements.txt
asgiref==3.8.1
Django==4.0.6
django-environ==0.11.2
djangorestframework==3.13.1
pillow==11.0.0
psycopg2-binary==2.9.10
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2
